[RTC-543] Request routing first iteration
