+++ lib/fishjam/room/id.ex
@@ -8,20 +8,17 @@ defmodule Fishjam.Room.ID do
   @doc """
   Based on the Room ID determines to which node it belongs to.
   Returns an error if the node isn't present in the cluster.
-
-  DISCLAIMER:
-    It should be used only with room_ids generated by generate/0, otherwise it can raise.
   """
   @spec determine_node(id()) ::
-          {:ok, node()} | {:error, :invalid_room_id} | {:error, :invalid_node}
+          {:ok, node()} | {:error, :invalid_room_id | :node_not_found}
   def determine_node(room_id) do
     with {:ok, room_id} <- validate_room_id(room_id),
-         node_name <- decode_node_name(room_id),
+         {:ok, node_name} <- decode_node_name(room_id),
          true <- node_present_in_cluster?(node_name) do
-      {:ok, node_name}
+      {:ok, String.to_atom(node_name)}
     else
       {:error, :invalid_room_id} -> {:error, :invalid_room_id}
-      false -> {:error, :invalid_node}
+      false -> {:error, :node_not_found}
     end
   end
 
@@ -44,16 +41,15 @@ defmodule Fishjam.Room.ID do
   end
 
   @doc """
-  Depending on feature flag "custom_room_name_disabled"
-    - uses `generate/0` to generate room_id
-    or
-    - parses the `room_id` provided by the client
+  Depending on feature flag "request_routing_enabled":
+    - if `true`, uses `generate/0` to generate room_id
+    - if `false`, parses the `room_id` provided by the client
   """
   @spec generate(nil | String.t()) :: {:ok, id()} | {:error, :invalid_room_id}
   def generate(nil), do: generate(UUID.uuid4())
 
   def generate(room_id) do
-    if Fishjam.FeatureFlags.custom_room_name_disabled?() do
+    if Fishjam.FeatureFlags.request_routing_enabled?() do
       {:ok, generate()}
     else
       validate_room_id(room_id)
@@ -65,18 +61,21 @@ defmodule Fishjam.Room.ID do
     |> String.split("-")
     |> Enum.take(-1)
     |> Enum.at(0)
-    |> Base.decode16!(case: :lower)
-    |> String.to_existing_atom()
+    |> Base.decode16(case: :lower)
+    |> case do
+      {:ok, node_name} -> {:ok, node_name}
+      :error -> {:error, :invalid_room_id}
+    end
   end
 
-  defp encoded_node_name do
+  defp encoded_node_name() do
     Node.self()
     |> Atom.to_string()
     |> Base.encode16(case: :lower)
   end
 
-  defp node_present_in_cluster?(node) do
-    node in [Node.self() | Node.list()]
+  defp node_present_in_cluster?(node_name) do
+    node_name in Enum.map([Node.self() | Node.list()], &Atom.to_string/1)
   end
 
   defp validate_room_id(room_id) when is_binary(room_id) do
