+++ lib/fishjam_web/controllers/recording_controller.ex
@@ -4,7 +4,8 @@ defmodule FishjamWeb.RecordingController do
 
   require Logger
 
-  alias Fishjam.Component.HLS.{Recording, RequestHandler}
+  alias Fishjam.Component.HLS.Recording
+  alias Fishjam.Component.HLS.Local.RequestHandler
   alias FishjamWeb.ApiSpec
 
   alias Plug.Conn
