+++ test/fishjam_web/controllers/component/hls_component_test.exs
@@ -4,7 +4,7 @@ defmodule FishjamWeb.Component.HlsComponentTest do
 
   import Mox
 
-  alias Fishjam.RoomService
+  alias Fishjam.Local.RoomService
 
   alias Fishjam.Component.HLS
